Loongarch opts
optimize loongarch64 inference

passed STC:
 LLR: 2.94 (-2.94,2.94) <0.00,2.00>
Total: 14208 W: 3792 L: 3510 D: 6906
Ptnml(0-2): 54, 1485, 3765, 1725, 75
https://tests.stockfishchess.org/tests/view/6a0a2a956524d21ee79b85a8

closes https://github.com/official-stockfish/Stockfish/pull/6832

No functional change

+++ src/nnue/nnue_accumulator.cpp
@@ -527,8 +527,8 @@ Bitboard get_changed_pieces(const std::array<Piece, SQUARE_NB>& oldPieces,
         const __m256i new_v = __lasx_xvld(reinterpret_cast<const void*>(&newPieces[i]), 0);
         const __m256i diff  = __lasx_xvxor_v(old_v, new_v);
         const __m256i mask  = __lasx_xvmsknz_b(diff);
-        const auto    lo    = static_cast<std::uint16_t>(__lasx_xvpickve2gr_d(mask, 0));
-        const auto    hi    = static_cast<std::uint16_t>(__lasx_xvpickve2gr_d(mask, 2));
+        const auto    lo    = __lasx_xvpickve2gr_d(mask, 0);
+        const auto    hi    = __lasx_xvpickve2gr_d(mask, 2);
 
         changed |= (static_cast<Bitboard>(lo) | (static_cast<Bitboard>(hi) << 16)) << i;
     }
@@ -546,8 +546,7 @@ Bitboard get_changed_pieces(const std::array<Piece, SQUARE_NB>& oldPieces,
         const __m128i diff  = __lsx_vxor_v(old_v, new_v);
         const __m128i mask  = __lsx_vmsknz_b(diff);
 
-        changed |= static_cast<Bitboard>(static_cast<std::uint16_t>(__lsx_vpickve2gr_d(mask, 0)))
-                << i;
+        changed |= static_cast<Bitboard>(__lsx_vpickve2gr_d(mask, 0)) << i;
     }
 
     return changed;
